layered architecture
