+++ MovieSite/module-application/src/main/java/com/root/moduleapplication/ModuleApplication.java
@@ -1,0 +1,13 @@
+package com.root.moduleapplication;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ModuleApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ModuleApplication.class, args);
+	}
+
+}
