+++ MovieSite/module-application/src/main/java/com/root/moduleapplication/position/service/PositionService.java
@@ -1,0 +1,34 @@
+package com.root.moduleapplication.position.service;
+
+import jakarta.transaction.Transactional;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+@RequiredArgsConstructor
+@Service
+public class PositionService {
+    private final PositionRepository positionRepository;
+
+    @Transactional
+    public List<Position> findAll() {
+        return positionRepository.findAll();
+    }
+
+    @Transactional
+    public Position save(Position position) {
+        return positionRepository.save(position);
+    }
+
+    @Transactional
+    public Position findById(Long id) {
+        return positionRepository.findById(id).orElse(null);
+    }
+
+    @Transactional
+    public void delete(Position position) {
+        positionRepository.delete(position);
+    }
+}
+
