+++ MovieSite/module-application/src/test/java/com/root/moduleapplication/ModuleApplicationTests.java
@@ -1,0 +1,13 @@
+package com.root.moduleapplication;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ModuleApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
