+++ MovieSite/module-presentation/src/main/java/com/root/modulepresentation/ModulePresentationApplication.java
@@ -1,0 +1,13 @@
+package com.root.modulepresentation;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ModulePresentationApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(ModulePresentationApplication.class, args);
+    }
+
+}
