+++ MovieSite/module-presentation/src/main/java/com/root/modulepresentation/position/dto/PositionRequestDto.java
@@ -1,0 +1,17 @@
+package com.root.modulepresentation.position.dto;
+
+import com.root.modulepresentation.movie.dto.MovieRequestDto;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class PositionRequestDto {
+    private MovieRequestDto movie;
+    private String seatNumber;
+}
