+++ MovieSite/module-presentation/src/test/java/com/root/modulepresentation/ModulePresentationApplicationTests.java
@@ -1,0 +1,13 @@
+package com.root.modulepresentation;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ModulePresentationApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
