+++ MovieSite/settings.gradle
@@ -1,1 +1,5 @@
 rootProject.name = 'MovieSite'
+include 'module-common'
+include 'module-application'
+include 'module-domain'
+include 'module-presentation'
