--- MovieSite/src/main/java/com/root/moviesite/MovieSiteApplication.java
@@ -1,13 +1,0 @@
-package com.root.moviesite;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class MovieSiteApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(MovieSiteApplication.class, args);
-    }
-
-}
