--- MovieSite/src/main/java/com/root/moviesite/movie/MovieRequestDto.java
@@ -1,6 +1,0 @@
-package com.root.moviesite.movie;
-
-public class MovieRequestDto {
-
-
-}
