--- MovieSite/src/main/java/com/root/moviesite/position/PositionRequestDto.java
@@ -1,4 +1,0 @@
-package com.root.moviesite.position;
-
-public class PositionRequestDto {
-}
