--- MovieSite/src/main/java/com/root/moviesite/user/UserRepository.java
@@ -1,7 +1,0 @@
-package com.root.moviesite.user;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-
-}
