--- MovieSite/src/main/java/com/root/moviesite/user/UserRequestDto.java
@@ -1,4 +1,0 @@
-package com.root.moviesite.user;
-
-public class UserRequestDto {
-}
