--- MovieSite/src/main/java/com/root/moviesite/user/UserService.java
@@ -1,34 +1,0 @@
-package com.root.moviesite.user;
-
-import jakarta.transaction.Transactional;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-import java.util.List;
-
-@RequiredArgsConstructor
-@Service
-public class UserService {
-
-    private final UserRepository userRepository;
-
-    @Transactional
-    public List<User> findAll() {
-        return userRepository.findAll();
-    }
-
-    @Transactional
-    public User findById(Long id) {
-        return userRepository.findById(id).orElse(null);
-    }
-
-    @Transactional
-    public User save(User user) {
-        return userRepository.save(user);
-    }
-
-    @Transactional
-    public void delete(User user) {
-        userRepository.delete(user);
-    }
-}
