--- MovieSite/src/main/resources/application.properties
@@ -1,2 +1,0 @@
-spring.application.name=MovieSite
-spring.jpa.hibernate.ddl-auto=create
