--- MovieSite/src/main/resources/application.yml
@@ -1,5 +1,0 @@
-spring:
-  jpa:
-    hibernate:
-      ddl-auto: create
-    # show-sql: true # DDL 출력
