--- MovieSite/src/test/java/com/root/moviesite/MovieSiteApplicationTests.java
@@ -1,13 +1,0 @@
-package com.root.moviesite;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class MovieSiteApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
